Rename indexed_modules.hpp -> indexed_modules.h
